Added utility scripts for raspberrypi configz.
A set of scripts that can be used to set up a raspberry pie. The goal
is to install and setup an NHS as fast as possible with as little
effort as possible. I’ve needed this a couple of times when my raspi
crashed miserably several times due to power outages.